Extract tick related items from component
